Rename File to Artifact to prepare a new object model around File entities
