+++ src/Source/DiskSources.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Haspadar\Piqule\Source;
 
 use FilesystemIterator;
-use Haspadar\Piqule\File\DiskFile;
+use Haspadar\Piqule\Artifact\DiskFile;
 use Haspadar\Piqule\PiquleException;
 use Override;
 use RecursiveDirectoryIterator;
