+++ src/Source/Source.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Haspadar\Piqule\Source;
 
-use Haspadar\Piqule\File\File;
+use Haspadar\Piqule\Artifact\File;
 
 final readonly class Source
 {
